+++ lib/generators/chaltron/install_generator.rb
@@ -127,31 +127,7 @@ JS
     end
 
     def setup_ajax_datatables
-      ajax_datatables_rails_file = 'config/initializers/ajax_datatables_rails.rb'
-      copy_file ajax_datatables_rails_file
-
-      # setup ajax-datatables-rails
-      db_adapter =
-        case ActiveRecord::Base.connection.adapter_name
-        when 'Mysql2'     then :mysql2
-        when 'SQLite'     then :sqlite3
-        when 'PostgreSQL' then :pg
-        else nil
-        end
-
-      if db_adapter.nil?
-        message =<<EOF
-  PAY ATTENTION!
-  ajax-datatables-rails gem (needed by chaltron) does not support #{options[:database]}.
-  See https://github.com/antillas21/ajax-datatables-rails#searching-on-non-text-based-columns.
-  You may experience problems!
-EOF
-        say message
-      else
-        gsub_file ajax_datatables_rails_file,
-          /# config.db_adapter = :mysql2/, "config.db_adapter = :#{db_adapter}"
-      end
-
+      copy_file 'config/initializers/ajax_datatables_rails.rb'
     end
 
     private