+++ lib/generators/chaltron/templates/config/initializers/ajax_datatables_rails.rb
@@ -1,7 +1,7 @@
 AjaxDatatablesRails.configure do |config|
-  # available options for db_adapter are: :pg, :mysql2, :sqlite3
-  # config.db_adapter = :mysql2
+  # available options for db_adapter are: :pg, :mysql, :mysql2, :sqlite, :sqlite3
+  # config.db_adapter = :pg
 
-  # available options for paginator are: :simple_paginator, :kaminari, :will_paginate
-  # config.paginator = :simple_paginator
+  # Or you can use your rails environment adapter if you want a generic dev and production
+  config.db_adapter = Rails.configuration.database_configuration[Rails.env]['adapter'].to_sym
 end