+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 If you are using bootstrap v3, refer to version 0, on [bootstrap3 branch](https://github.com/vicvega/chaltron/blob/bootstrap3/CHANGELOG.md).
 
 ### 2.0.1 - current version - unreleased
+* fix issue with install generator when database password is needed
 
 ### 2.0.0 - 2020-05-25
 * refactor all assets dependencies